Set default branch for repository on init
